Move collector.Registry into the internal/collector package.
