+++ internal/collector/registry.go
